+++ src/samples/middlewares.js
@@ -2,44 +2,12 @@
   Global Middlewares
   These middlewares will be added to start of the the express app 
 */
-const _globals = [
+const globals = [
   (req, res, next) => {
     console.log(req.originalUrl);
     next();
   }
 ]
-
-/* 
-  Used in VS Code Mock Server extension
-  This method is called only on generating db suing MockServer: Generate Db Command
-  It will be called for each entry/hits in a HAR/Kibana formatted data
-  Here you can return your custom route and routeConfig
-  `_harEntryCallback`, `_kibanaHitsCallback` is a reserved word for generating Db 
-*/
-const _harEntryCallback = (entry, routePath, routeConfig) => {
-  // your code goes here ...
-  return { [routePath]: routeConfig }
-};
-const _kibanaHitsCallback = (hit, routePath, routeConfig) => {
-  // your code goes here ...
-  return { [routePath]: routeConfig }
-};
-
-/* 
-  Used in VS Code Mock Server extension
-  This method is called only on generating db suing MockServer: Generate Db Command
-  It will be called at last of all entry/hits looping.
-  Here you can return your custom db
-  `_harDbCallback`, `_KibanaDbCallback` is a reserved word for generating Db
-*/
-const _harDbCallback = (data, db) => {
-  // your code goes here ...
-  return db;
-};
-const _KibanaDbCallback = (data, db) => {
-  // your code goes here ...
-  return db;
-};
 
 /* 
   This is a Express middleware used to call on a specific routes.
@@ -82,11 +50,7 @@ module.exports = (mockServer) => {
   // Your Global middleware logic here before setting default middlewares by the MockServer
 
   return {
-    _globals,
-    _harEntryCallback,
-    _kibanaHitsCallback,
-    _harDbCallback,
-    _KibanaDbCallback,
+    globals,
     DataWrapper,
     CustomLog,
     GetStoreValue,