+++ src/server/middlewares/initialMiddleware.ts
@@ -4,10 +4,10 @@ import { Locals } from '../types/common.types';
 import * as ValidTypes from '../types/valid.types';
 import { setRequestUrl } from './fetch';
 
-export default (routePath: string, config: ValidTypes.Config, getDb: () => ValidTypes.Db, getStore: () => ValidTypes.Store) => {
+export default (routePath: string, config: ValidTypes.Config, getDb: (routePath?: string | string[]) => ValidTypes.RouteConfig | ValidTypes.Db, getStore: () => ValidTypes.Store) => {
   return async (req: express.Request, res: express.Response, next: express.NextFunction) => {
     try {
-      const routeConfig = getDb()[routePath] || {} as ValidTypes.RouteConfig;
+      const routeConfig = (getDb(routePath) || {}) as ValidTypes.RouteConfig;
       routeConfig.store && !_.isPlainObject(routeConfig.store) && (routeConfig.store = {});
 
       const locals = res.locals as Locals