+++ tests/mock/middlewares/middlewares.js
@@ -1,6 +1,6 @@
 module.exports = {
   logger: (req, res, next) => { next() },
   auth: [(req, res, next) => { next() }, (req, res, next) => { next() }],
-  _globals: [(req, res, next) => { next() }, {}, "XXX"],
+  globals: [(req, res, next) => { next() }, {}, "XXX"],
   dummy: {}
 }